I think I've finally fixed the timing problems with Login2Me.
I've moved some sleeps around and adjusted some values. I realized
that I was creating a new browser each time the scraper function
was called, but the timeout on that browser session wasn't getting
cleaned up fast enough on the Selenium side. This can be remedied
by setting a timeout value of something like 20-25 when running
the selenium server thing.